Rename solutionOne file to maintain the same convention
