+++ src/logic/solutionOne.ts
